services/backend/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
sqlmodel==0.0.8
python-dotenv==1.0.0
pyjwt==2.8.0
g4f==0.1.4.2
sse-starlette==1.6.5
starlette==0.27.0
starlette-context==0.3.6
oss2==2.18.3
python-multipart==0.0.6
pydantic-settings==2.0.3
mysqlclient==2.2.0